+++ tests/test_weather.py
@@ -8,6 +8,7 @@
 
 from __future__ import annotations
 
+import datetime
 import json
 import pathlib
 import time
@@ -42,18 +43,34 @@ async def test_get_weather_data(mock_get_from_lco_api, start_time: str | float):
     assert data.height == 94
 
 
+@pytest.mark.parametrize(
+    "now", [None, 1732680854.704, datetime.datetime(2024, 11, 27, 4, 14, 14, 704000)]
+)
 async def test_is_weather_data_safe(
     mock_get_from_lco_api,
     mocker: pytest_mock.MockerFixture,
+    now: float | datetime.datetime | None,
 ):
     mocker.patch.object(time, "time", return_value=1732680854.704)
 
     data = await get_weather_data("2024-11-27T03:56:10.618329")
 
-    assert is_weather_data_safe(data, "wind_speed_avg", 35)
-    assert not is_weather_data_safe(data, "wind_speed_avg", 5)
+    assert is_weather_data_safe(data, "wind_speed_avg", 35, now=now)
+    assert not is_weather_data_safe(data, "wind_speed_avg", 5, now=now)
 
     # Values are such that the max avg wind speed in the last 60 minutes is 12.11
     # and the maximum in the last 30 minutes is 10.18 mph.
-    assert is_weather_data_safe(data, "wind_speed_avg", 12.5, reopen_value=10)
-    assert not is_weather_data_safe(data, "wind_speed_avg", 12, reopen_value=10)
+    assert is_weather_data_safe(
+        data,
+        "wind_speed_avg",
+        12.5,
+        reopen_value=10,
+        now=now,
+    )
+    assert not is_weather_data_safe(
+        data,
+        "wind_speed_avg",
+        12,
+        reopen_value=10,
+        now=now,
+    )